Move defaults to tfbridge and move strategies to tfbridge/tokens
