--- pkg/tfbridge/defaults/package.go
@@ -1,5 +1,0 @@
-// Experimental extensions to the tfbridge module.
-//
-// API exposed by this package is unstable and subject to change. As it stabilizes it will be moved to the parent
-// package tfbridge.
-package defaults
