+++ pkg/tfbridge/tokens/tokens.go
@@ -1,0 +1,54 @@
+// Copyright 2016-2023, Pulumi Corporation.
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+package tokens
+
+import (
+	"fmt"
+	"unicode"
+
+	"github.com/pulumi/pulumi/pkg/v3/codegen/cgstrings"
+)
+
+const (
+	defaultMinimumModuleSize    = 5
+	defaultMinimumSubmoduleSize = 5
+)
+
+// A function that joins a module and name into a pulumi type token.
+//
+// For example:
+//
+//	func(module, name string) (string, error) {
+//	 return fmt.Sprintf("pkgName:%s:%s", module, name), nil
+//	}
+type Make func(module, name string) (string, error)
+
+// Convert a Terraform token to a Pulumi token with the standard mapping.
+//
+// The mapping is
+//
+//	(pkg, module, name) => pkg:module/lowerFirst(name):name
+func MakeStandard(pkgName string) Make {
+	return func(module, name string) (string, error) {
+		lowerName := string(unicode.ToLower(rune(name[0]))) + name[1:]
+		return fmt.Sprintf("%s:%s/%s:%s", pkgName, module, lowerName, name), nil
+	}
+}
+
+func upperCamelCase(s string) string { return cgstrings.UppercaseFirst(camelCase(s)) }
+
+func camelCase(s string) string {
+	return cgstrings.ModifyStringAroundDelimeter(s, "_", cgstrings.UppercaseFirst)
+}
