+++ pkg/tfbridge/x/package.go
@@ -3,5 +3,7 @@
 // API exposed by this package is unstable and subject to change. As it stabilizes it will be moved to the parent
 // package tfbridge.
 //
-// Deprecated: This package has been moved to "github.com/pulumi/pulumi-terraform-bridge/v3/pkg/tfbridge/defaults"
+// Deprecated: This package has been moved to
+// [github.com/pulumi/pulumi-terraform-bridge/v3/pkg/tfbridge] and
+// [github.com/pulumi/pulumi-terraform-bridge/v3/pkg/tfbridge/tokens]
 package x
