+++ pkg/tfbridge/x/strategy.go
@@ -21,14 +21,14 @@ import (
 // A generic remapping strategy.
 //
 // Deprecated: This item has been moved to
-// "github.com/pulumi/pulumi-terraform-bridge/v3/pkg/tfbridge/defaults.ElementStrategy"
+// [github.com/pulumi/pulumi-terraform-bridge/v3/pkg/tfbridge.ElementStrategy]
 type Strategy[T tfbridge.ResourceInfo | tfbridge.DataSourceInfo] func(tfToken string) (*T, error)
 
 // Describe the mapping from resource and datasource tokens to Pulumi resources and
 // datasources.
 //
 // Deprecated: This item has been moved to
-// "github.com/pulumi/pulumi-terraform-bridge/v3/pkg/tfbridge/defaults.Strategy"
+// [github.com/pulumi/pulumi-terraform-bridge/v3/pkg/tfbridge.Strategy]
 type DefaultStrategy struct {
 	Resource   ResourceStrategy
 	DataSource DataSourceStrategy
@@ -37,11 +37,11 @@ type DefaultStrategy struct {
 // A strategy for generating missing resources.
 //
 // Deprecated: This item has been moved to
-// "github.com/pulumi/pulumi-terraform-bridge/v3/pkg/tfbridge/defaults.ResourceStrategy"
+// [github.com/pulumi/pulumi-terraform-bridge/v3/pkg/tfbridge.ResourceStrategy]
 type ResourceStrategy = Strategy[tfbridge.ResourceInfo]
 
 // A strategy for generating missing datasources.
 //
 // Deprecated: This item has been moved to
-// "github.com/pulumi/pulumi-terraform-bridge/v3/pkg/tfbridge/defaults.DataSourceStrategy"
+// [github.com/pulumi/pulumi-terraform-bridge/v3/pkg/tfbridge.DataSourceStrategy]
 type DataSourceStrategy = Strategy[tfbridge.DataSourceInfo]
