+++ pkg/tfbridge/x/unmappable.go
@@ -22,7 +22,7 @@ import (
 // Indicate that a token cannot be mapped.
 //
 // Deprecated: This item has been moved to
-// "github.com/pulumi/pulumi-terraform-bridge/v3/pkg/tfbridge/defaults.UnmappableError"
+// [github.com/pulumi/pulumi-terraform-bridge/v3/pkg/tfbridge/UnmappableError]
 type UnmappableError struct {
 	TfToken string
 	Reason  error
